[Cellular Lib CBMC] CBMC for cellular_common.h and cellular_common_api.h (#12)
* [Cellular Lib CBMC] CBMC for cellular_common.h and cellular_common_api.h

[Cellular Library CBMC] setup build environment

[Cellular Library CBMC] CBMC for Cellular_CommonInit

[Cellular Library CBMC] CBMC for Cellular_CommonClean

[Cellular Library CBMC] CBMC for Cellular_CommonATCommandRaw

[Cellular Library CBMC] CBMC for Cellular_CommonCreateSocket

[Cellular Library CBMC] CBMC for Cellular_CommonSocketSetSockOpt

[Cellular Library CBMC] modification fo Cellular_CommonATCommandRaw CBMC for more function coverage

[Cellular Library CBMC] modification fo Cellular_CommonCleanup CBMC for more function coverage

[Cellular Library CBMC] modification fo Cellular_CommonCreateSocket CBMC for more function coverage

[Cellular Library CBMC] modification fo Cellular_CommonInit CBMC for more function coverage

[Cellular Library CBMC] modification fo Cellular_CommonSocketSetSockOpt CBMC for more function coverage

[Cellular Library CBMC] modification for Cellular_CommonATCommandRaw CBMC for more function coverage

[Cellular Library CBMC]  coverage for Cellular_CommonSocketSetSockOpt

[Cellular Library CBMC] CBMC for Cellular_CommonRfOn

[Cellular Library CBMC] CBMC for Cellular_CommonRfOff

[Cellular Library CBMC] CBMC for Cellular_CommonGetIPAddress

[Cellular Library CBMC] CBMC for Cellular_CommonGetModemInfo

[Cellular Library CBMC] CBMC for Cellular_CommonGetEidrxSettings

[Cellular Library CBMC] CBMC for Cellular_CommonSetEidrxSettings

[Cellular Library CBMC] CBMC for Cellular_CommonGetRegisteredNetwork

[Cellular Library CBMC] CBMC for Cellular_CommonGetNetworkTime

[Cellular Library CBMC] CBMC for Cellular_CommonGetServiceStatus

[Cellular Library CBMC] CBMC for Cellular_CommonSetPdnConfig

[Cellular Library CBMC] CBMC for Cellular_CommonGetPsmSettings

[Cellular Library CBMC] CBMC for Cellular_CommonSetPsmSettings

[Cellular Library CBMC] CBMC for Cellular_CommonGetSimCardInfo

[Cellular Library CBMC] CBMC for Cellular_CommonGetSimCardLockStatus

[Cellular Library CBMC] CBMC for Cellular_CommonUrcProcessCereg

[Cellular Library CBMC] CBMC for Cellular_CommonUrcProcessCgreg

[Cellular Library CBMC] CBMC for Cellular_CommonUrcProcessCreg

[Cellular Library CBMC] Use macro for unwindset instead of number

[Cellular Library CBMC] CBMC for Cellular_CommonRegisterUrcNetworkRegistrationEventCallback

[Cellular Library CBMC] CBMC for Cellular_CommonRegisterUrcPdnEventCallback

[Cellular Library CBMC] CBMC for Cellular_CommonRegisterUrcSignalStrengthChangedCallback

[Cellular Library CBMC] CBMC for Cellular_CommonRegisterUrcGenericCallback

[Cellular Library CBMC] CBMC for Cellular_CommonRegisterModemEventCallback

[Cellular Library CBMC] CBMC for Cellular_CommonSocketRegisterDataReadyCallback

[Cellular Library CBMC] CBMC for Cellular_CommonSocketRegisterSocketOpenCallback

[Cellular Library CBMC] CBMC for Cellular_CommonSocketRegisterClosedCallback

[Cellular Library CBMC] Remove unused variable.

[Cellular Library CBMC] modify Makefile for unwind set configuration.

[Cellular Library CBMC] correct loop parameter to avoid unwind

[Cellular Library CBMC] CBMC for _Cellular_LibCleanup

[Cellular Library CBMC] CBMC for _Cellular_NetworkRegistrationCallback

[Cellular Library CBMC] CBMC for _Cellular_PdnEventCallback

[Cellular Library CBMC] CBMC for _Cellular_SignalStrengthChangedCallback

[Cellular Library CBMC] CBMC for _Cellular_GenericCallback

[Cellular Library CBMC] correct README.md

[Cellular Library CBMC] CBMC for _Cellular_ModemEventCallback

[Cellular Library CBMC] CBMC for _Cellular_TranslateAtCoreStatus

[Cellular Library CBMC] Adjust parameter to verify _Cellular_CreateSocketData

[Cellular Library CBMC] CBMC for _Cellular_RemoveSocketData

[Cellular Library CBMC] CBMC for _Cellular_IsValidSocket

[Cellular Library CBMC] CBMC for _Cellular_ConvertCsqSignalRssi

[Cellular Library CBMC] CBMC for _Cellular_ConvertCsqSignalBer

[Cellular Library CBMC] CBMC for _Cellular_GetModuleContext

[Cellular Library CBMC] CBMC for _Cellular_ComputeSignalBars

[Cellular Library CBMC] CBMC for _Cellular_GetCurrentRat

[Cellular Library CBMC] CBMC for _Cellular_GetSocketData

[Cellular Library CBMC] CBMC for _Cellular_TimeoutAtcmdDataSendRequestWithCallback

[Cellular Library CBMC] CBMC for _Cellular_TimeoutAtcmdDataRecvRequestWithCallback

[Cellular Library CBMC] CBMC for _Cellular_TimeoutAtcmdDataSendSuccessToken

[Cellular Library CBMC] correct coding covention

* [Cellular Lib CBMC] Fix missing error code.

* [Cellular Lib CBMC] Remove the internal API in Makefile

+++ common/include/cellular_common.h
@@ -50,7 +50,7 @@ typedef struct CellularAtReq
     const char * pAtRspPrefix;
     CellularATCommandResponseReceivedCallback_t respCallback;
     void * pData;
-    uint32_t dataLen;
+    uint16_t dataLen;
 } CellularAtReq_t;
 
 /**
